+++ qiskit_aqt_provider/aqt_resource.py
@@ -10,7 +10,6 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-import time
 import typing
 import warnings
 from dataclasses import dataclass
@@ -281,8 +280,6 @@ class AQTResource(_ResourceBase[AQTOptions]):
     def result(self, job_id: UUID) -> ResultResponse:
         """Query the result for a specific job.
 
-        If a transient error occurs during the query it is retried until a result is obtained.
-
         .. tip:: This is a low-level method. Use the
             :meth:`AQTJob.result <qiskit_aqt_provider.aqt_job.AQTJob.result>`
             method to retrieve the result of a job described by a
@@ -293,30 +290,9 @@ class AQTResource(_ResourceBase[AQTOptions]):
 
         Returns:
             AQT API payload with the job results.
-
-        Raises:
-            APIError: If a non transient error occurs.
         """
-        result = None
-        while result is None:
-            try:
-                resp = self._http_client.get(f"/result/{job_id}")
-                resp.raise_for_status()
-                result = ResultResponse.model_validate(resp.json())
-            except (httpx.RequestError, httpx.HTTPStatusError) as ex:  # noqa: PERF203
-                if isinstance(ex, httpx.HTTPStatusError) and resp.status_code not in (
-                    429,
-                    499,
-                    500,
-                    502,
-                    503,
-                    504,
-                ):
-                    http_response_raise_for_status(resp)
-                warnings.warn(f"Retrying to retrieve the job status after transient error {ex}.")
-                time.sleep(1)
-
-        return result
+        resp = http_response_raise_for_status(self._http_client.get(f"/result/{job_id}"))
+        return ResultResponse.model_validate(resp.json())
 
 
 class AQTDirectAccessResource(_ResourceBase[AQTDirectAccessOptions]):