+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Update aqt-connector to 0.3.0 (#251)
 * Unify user experience to acquire backends accessed via cloud or directly (#256)
 * Add hint to docs regarding token expiration, remove token from examples (#260)
-* Retry for transient errors in AQTResource result fetching (#274)
+* Retry for transient errors in AQTResource result polling (#274)
 
 ## qiskit-aqt-provider v1.14.0